fixed syntax error and created utility function for pagination data filter
